refactor(database): moved table create/drop/validate to new files
